+++ Datamigratie.Common/Services/Det/Models/DetZaak.cs
@@ -14,6 +14,7 @@ namespace Datamigratie.Common.Services.Det.Models
     {
         public string? AangemaaktDoor { get; set; }
         public string? Afdeling { get; set; }
+        public DetBetaalgegevens? Betaalgegevens { get; set; }
 
         [JsonConverter(typeof(DetZonedDateTimeConverter))]
         public DateTimeOffset CreatieDatumTijd { get; set; }
@@ -41,6 +42,10 @@ namespace Datamigratie.Common.Services.Det.Models
         public required List<DetDocumentVersie> DocumentVersies { get; set; }
         public string? Kenmerk { get; set; }
         public required string Titel { get; set; }
+        public string? Beschrijving {get; set; }
+        public DocumentTaal? Taal { get; set; }
+
+        public DocumentVorm? DocumentVorm { get; set; }
     }
 
     public class DetDocumentVersie
@@ -48,8 +53,26 @@ namespace Datamigratie.Common.Services.Det.Models
         public required int Versienummer { get; set; }
         public required long DocumentInhoudID { get; set; }
         public required string Bestandsnaam { get; set; }
-        public required string Mimetype { get; set; }
+        public string Mimetype { get; set; }
         public long? Documentgrootte { get; set; }
+        public string? Auteur { get; set; }
         public required DateOnly Creatiedatum { get; set; }
+    }
+
+    public class DetBetaalgegevens
+    {
+        public decimal? Bedrag { get; set; }
+        public string? Betaalstatus { get; set; }
+        public string? Kenmerk { get; set; }
+    }
+
+    public class DocumentTaal
+    {
+        public required string FunctioneelId { get; set; }
+    }
+
+    public class DocumentVorm
+    {
+        public required string Naam { get; set; }
     }
 }